feat: manual-only score entry, admin avatar editing, gender RPC type fix
- Score sheet: hide the number-input spinner arrows (appearance:textfield plus
  the webkit spin-button reset) so scores are typed in directly. The field stays
  type=number for the numeric keypad on mobile.
- Admins can replace a player's photo. The edit sheet gains an avatar preview and
  a "เปลี่ยนรูปโปรไฟล์" picker; on save the new file is uploaded and set.
  Migration 011 adds admin_update_player_avatar, which validates the new object
  exactly like registration (belongs to this project, exists in storage, not
  already used) before updating players.avatar_url. The local demo compresses and
  stores the image inline, mirroring registration.
- Fix admin_set_player_gender in migration 010: it declared demo_slot as integer
  while the column is smallint, so Postgres raised "structure of query does not
  match function result type" the moment an admin set a gender. Both 010 and 011
  now return demo_slot as smallint.

Verified in the local demo (score field has no spinner, edit sheet offers a photo
picker); tests 56/56, lint, build green.

+++ supabase/migrations/010_gender_divisions.sql
@@ -19,7 +19,7 @@ create unique index if not exists bracket_matches_division_slot_unique
   on public.bracket_matches (draw_version, division, round_number, match_position);
 
 create or replace function public.admin_set_player_gender(p_public_id text, p_gender text)
-returns table (public_id text, nickname text, department text, email text, avatar_url text, registered_at timestamptz, gender text, is_demo boolean, demo_slot integer)
+returns table (public_id text, nickname text, department text, email text, avatar_url text, registered_at timestamptz, gender text, is_demo boolean, demo_slot smallint)
 language plpgsql security definer set search_path = pg_catalog
 as $$
 declare current_state public.tournament_state%rowtype;
